requirements.txt
langchain-community>=0.0.17
langchain-huggingface>=0.0.1
sentence-transformers
faiss-cpu
streamlit
pypdf
openai
python-dotenv

openai>=1.3.5
langchain-openai>=0.1.0
